Rename "job" to "pod" & drop instances
On reflection, conflating a definition of a pod with how many
instances to run globally isn't good.  Ideally, a change to a pod
definition triggers a rolling restart across the cluster - which isn't
really doable if it *also* tracks the number of instances, and so
scaling up or down would be picked up as a change.

Rather than implement some sort of diffing logic, to only trigger
restarts on meaningful changes, just split the pod and job apart.

I think this is essentially the Pod / Deployment difference in k8s.